backend/requirements.txt
# FastAPI Backend Requirements
# Install using: pip install -r requirements.txt

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning
torch==2.0.1
transformers==4.30.0

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0

# Data Processing
pandas==2.1.3
numpy==1.24.3

# HTTP Requests
requests==2.31.0
beautifulsoup4==4.12.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0

# Optional (for production)
gunicorn==21.2.0